test(dashboard): add DragDroppable mock and ChartHolder fullscreen tests
