+++ superset-frontend/src/dashboard/components/gridComponents/ChartHolder/ChartHolder.test.tsx
@@ -42,57 +42,7 @@ import {
 import ChartHolder, { CHART_MARGIN } from './ChartHolder';
 import { GRID_BASE_UNIT, GRID_GUTTER_SIZE } from '../../../util/constants';
 
-jest.mock('src/dashboard/components/dnd/DragDroppable', () => {
-  // eslint-disable-next-line @typescript-eslint/no-var-requires -- jest.mock factory
-  const React = require('react');
-  const actual = jest.requireActual(
-    'src/dashboard/components/dnd/DragDroppable',
-  );
-  // eslint-disable-next-line @typescript-eslint/no-var-requires -- jest.mock factory
-  const cx = require('classnames');
-
-  function MockDraggable(props: Record<string, unknown>) {
-    const { editMode, orientation, children, disableDragDrop } = props as {
-      editMode?: boolean;
-      orientation?: 'row' | 'column';
-      disableDragDrop?: boolean;
-      children: (childProps: Record<string, unknown>) => unknown;
-    };
-    const childFn = children as (
-      childProps: Record<string, unknown>,
-    ) => unknown;
-    const childProps = editMode
-      ? {
-          dragSourceRef: jest.fn(),
-          dropIndicatorProps: null,
-          draggingTabOnTab: false,
-          'data-test': 'dragdroppable-content',
-        }
-      : {
-          dropIndicatorProps: null,
-          'data-test': 'dragdroppable-content',
-        };
-    return React.createElement(
-      'div',
-      {
-        'data-test': 'dragdroppable-object',
-        'data-disable-drag-drop': String(!!disableDragDrop),
-        className: cx(
-          'dragdroppable',
-          editMode && 'dragdroppable--edit-mode',
-          orientation === 'row' && 'dragdroppable-row',
-          orientation === 'column' && 'dragdroppable-column',
-        ),
-      },
-      childFn(childProps),
-    );
-  }
-
-  return {
-    ...actual,
-    Draggable: MockDraggable,
-  };
-});
+jest.mock('src/dashboard/components/dnd/DragDroppable');
 
 const DEFAULT_HEADER_HEIGHT = 22;
